requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
cosine-annealing-warmup==2.0
einops==0.6.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
importlib-metadata==6.0.0
lightning-utilities==0.6.0.post0
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wget==3.2
yarl==1.8.2
zipp==3.13.0
